+++ Boards/Templates/Baremetal/CANFD_Ext_Loopback.c
@@ -22,6 +22,7 @@
 
 #include <stdio.h>
 #include <string.h>
+#include <inttypes.h>
 #include <RTE_Components.h>
 #include CMSIS_device_header
 #include "pinconf.h"
@@ -256,7 +257,7 @@ static void canfd_lbe_demo_thread_entry(void)
                                               &service_error_code);
     if(error_code)
     {
-        printf("SE Error: HFOSC clk enable = %d\n", (int)error_code);
+        printf("SE Error: HFOSC clk enable = %"PRIu32"\n", error_code);
         return;
     }
 
@@ -267,7 +268,7 @@ static void canfd_lbe_demo_thread_entry(void)
                                               &service_error_code);
     if(error_code)
     {
-        printf("SE Error: 160 MHz clk enable = %d\n", (int)error_code);
+        printf("SE Error: 160 MHz clk enable = %"PRIu32"\n", error_code);
         return;
     }
 
@@ -278,7 +279,7 @@ static void canfd_lbe_demo_thread_entry(void)
     ret_val = board_pins_config();
     if (ret_val != 0)
     {
-        printf("Error in pin-mux configuration: %d\n", ret_val);
+        printf("Error in pin-mux configuration: %"PRId32"\n", ret_val);
         return;
     }
 
@@ -290,14 +291,14 @@ static void canfd_lbe_demo_thread_entry(void)
     ret_val = board_canfd_pins_config();
     if(ret_val != 0)
     {
-        printf("Error in pin-mux configuration: %d\n", ret_val);
+        printf("Error in pin-mux configuration: %"PRId32"\n", ret_val);
         return;
     }
 #endif
 
     /* Get CANFD capabilities */
     can_capabilities = CANFD_instance->GetCapabilities();
-    printf("Num of objects supported: %d\r\n", can_capabilities.num_objects);
+    printf("Num of objects supported: %"PRIu8"\r\n", can_capabilities.num_objects);
 
     /* Initializing CANFD Access struct */
     ret_val = CANFD_instance->Initialize(cb_unit_event, cb_object_event);
@@ -465,7 +466,7 @@ uninitialise_canfd:
                                               &service_error_code);
     if(error_code)
     {
-        printf("SE Error: HFOSC clk disable = %d\n", (int)error_code);
+        printf("SE Error: HFOSC clk disable = %"PRIu32"\n", error_code);
         return;
     }
     /* Disables the 160MHz clock */
@@ -475,7 +476,7 @@ uninitialise_canfd:
                                               &service_error_code);
     if(error_code)
     {
-        printf("SE Error: 160 MHz clk disable = %d\n", (int)error_code);
+        printf("SE Error: 160 MHz clk disable = %"PRIu32"\n", error_code);
         return;
     }
 
@@ -527,7 +528,7 @@ static void canfd_check_error(void)
         {
             printf("Error: CANFD Bus OFF:\r\n");
         }
-        printf("Error in CANFD-->Error Code:%d\r\n", cur_sts.last_error_code);
+        printf("Error in CANFD-->Error Code:%"PRIu8"\r\n", cur_sts.last_error_code);
 
         bus_error = false;
     }
@@ -556,8 +557,8 @@ static bool canfd_process_rx_message(void)
          * performs the below operations */
         if(rx_msg_header.rtr == 1U)
         {
-            printf("Rx msg:\r\n    Type:Remote frame, Id:%lu",
-                   (rx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)));
+            printf("Rx msg:\r\n    Type:Remote frame, Id:%"PRIu32"",
+                   (uint32_t)(rx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)));
         }
         else
         {
@@ -577,8 +578,8 @@ static bool canfd_process_rx_message(void)
                         return msg_rx_sts;
                     }
 
-                    printf("Id:%lu, Len:%d:\r\n    Data:",
-                           (rx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)),
+                    printf("Id:%"PRIu32", Len:%"PRIu8":\r\n    Data:",
+                           (uint32_t)(rx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)),
                             rx_msg_size);
                     for(iter = 0; iter < rx_msg_size; iter++)
                     {
@@ -701,8 +702,8 @@ static void canfd_transmit_message(const CANFD_FRAME msg_type)
     if(status == ARM_DRIVER_OK)
     {
         rx_msg_size = tx_msg_size;
-        printf("Tx Msg:\r\n    Id:%lu, Len:%d: \r\n    Data:",
-                (tx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)), tx_msg_size);
+        printf("Tx Msg:\r\n    Id:%"PRIu32", Len:%"PRIu8": \r\n    Data:",
+                (uint32_t)(tx_msg_header.id & (~ARM_CAN_ID_IDE_Msk)), tx_msg_size);
         for(iter = 0; iter < tx_msg_size; iter++)
         {
             printf("%c", tx_data[iter]);
